+++ packages/database/drizzle/maimai/0005_add-column-release-intl-to-version-data.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "music_version" ADD COLUMN "version_intl" text;